fix(cf): sanitise service binding names
