+++ clouddriver-cloudfoundry/src/main/java/com/netflix/spinnaker/clouddriver/cloudfoundry/client/model/v2/CreateServiceBinding.java
@@ -16,13 +16,13 @@
 
 package com.netflix.spinnaker.clouddriver.cloudfoundry.client.model.v2;
 
+import static com.netflix.spinnaker.clouddriver.cloudfoundry.client.CloudFoundryClientUtils.convertToValidServiceBindingName;
+
 import java.util.Map;
 import lombok.Getter;
 
 @Getter
 public class CreateServiceBinding {
-  private static final String VALID_BINDING_NAME_PATTERN = "[^A-Za-z0-9-]+";
-
   private final String serviceInstanceGuid;
   private final String appGuid;
   private final String name;
@@ -35,7 +35,7 @@ public class CreateServiceBinding {
       final Map<String, Object> parameters) {
     this.serviceInstanceGuid = serviceInstanceGuid;
     this.appGuid = appGuid;
-    this.name = sanitiseBindingName(name);
+    this.name = convertToValidServiceBindingName(name);
     this.parameters = parameters;
   }
 
@@ -43,10 +43,6 @@ public class CreateServiceBinding {
       final String serviceInstanceGuid, final String appGuid, final String name) {
     this.serviceInstanceGuid = serviceInstanceGuid;
     this.appGuid = appGuid;
-    this.name = sanitiseBindingName(name);
-  }
-
-  private String sanitiseBindingName(final String name) {
-    return name.replaceAll(VALID_BINDING_NAME_PATTERN, "-");
+    this.name = convertToValidServiceBindingName(name);
   }
 }
